Updated the SQL query for the foreign keys & rowcounts
The Foreign key table wasn't initally linking to the `information_schema.key_column_usage` which gives us the FK column name using a particular constraint. This is fixed and I've changed the columns being returned.

The rowcounts are super simple and they're just selecting the count for a particular column with a particular value.